tests/wpt/web-platform-tests/app-history/transition/fragment-nav.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    appHistory.onnavigateerror = t.unreached_func("navigateerror must not fire");
    appHistory.onnavigatesuccess = t.step_func_done(() => {
      assert_equals(appHistory.transition, null);
    });

    assert_equals(appHistory.transition, null);
    location.href = "#1";
    assert_equals(appHistory.transition, null);
  }, 0);
}, "non-intercepted fragment navigations leave transition null");
</script>
